Instalacao

+++ src/main/java/com/sunnymeter/api/core/instalacao/DadosDetalhesConsumoInstalacao.java
@@ -8,7 +8,7 @@ public record DadosDetalhesConsumoInstalacao (
 		int  	dia_referencia,
 		String 	mes_referencia,
 		int 	ano_referencia,
-		long	dias_para_acabar_o_mes,
+		int	dias_para_acabar_o_mes,
 		double	consumo_mensal_medio_kwh,
 		double	consumo_diario_medio_kwh,
 		double	consumo_mensal_estimado_kwh
